refactor: split status and condition styles out of styles.go
styles.go was at the 800-line revive cap. Move the status-severity
block into styles_status.go and the condition-polarity block into
styles_condition.go, pure relocation with no behavior change.
