+++ internal/ui/styles.go
@@ -4,11 +4,8 @@ import (
 	"image/color"
 	"strconv"
 	"strings"
-	"unicode"
 
 	"charm.land/lipgloss/v2"
-
-	"github.com/janosmiko/lfk/internal/model"
 )
 
 // Tokyonight Storm color palette default values. These back the mutable
@@ -474,312 +471,6 @@ func AgeStyle(age string) lipgloss.Style {
 		return lipgloss.NewStyle().Foreground(lipgloss.Color(ColorBorder)).Background(BaseBg)
 	default:
 		return DimStyle
-	}
-}
-
-// statusSev classifies a resource status string into a severity bucket. It is
-// the single source of truth shared by StatusStyle (which maps it to a color)
-// and StatusSeverityRank (which maps it to an ordering), so coloring and
-// severity ordering can never drift — and the ordering stays correct in
-// no-color mode, where the styles carry no distinguishing foreground.
-type statusSev int
-
-const (
-	sevUnknown     statusSev = iota // unknown non-empty status -> dimmed
-	sevBlank                        // "" -> neutral
-	sevDefault                      // literal "default" -> primary
-	sevNormal                       // "Normal" event type -> dim
-	sevRunning                      // healthy / ready / synced
-	sevDone                         // succeeded / completed
-	sevProgressing                  // pending / progressing / out-of-sync / warning
-	sevFailed                       // failed / degraded / errored
-)
-
-func statusSeverity(status string) statusSev {
-	switch status {
-	case "default":
-		return sevDefault
-	case "Running", "Active", "Bound", "Available", "Ready",
-		"Healthy", "Healthy/Synced", "Synced",
-		"Deployed",
-		"SecretSynced", "Created", "Updated", "Valid",
-		"Established":
-		return sevRunning
-	case "Succeeded", "Completed",
-		"Superseded":
-		return sevDone
-	case "Pending", "ContainerCreating", "PodInitializing", "Terminating",
-		"Waiting", "Init", "NotReady",
-		"Progressing", "Progressing/Synced", "Progressing/OutOfSync",
-		"Missing", "Suspended", "Unknown", "Reconciling",
-		"Healthy/OutOfSync", "Missing/OutOfSync", "Suspended/OutOfSync",
-		"OutOfSync",
-		"Pending-install", "Pending-upgrade", "Pending-rollback", "Uninstalling",
-		"Warning":
-		return sevProgressing
-	case "Normal":
-		return sevNormal
-	case "Failed", "CrashLoopBackOff", "Error", "ImagePullBackOff", "Terminated",
-		"Degraded", "Degraded/Synced", "Degraded/OutOfSync",
-		"Missing/Synced",
-		"OOMKilled", "ErrImagePull", "CreateContainerConfigError",
-		"SecretSyncedError", "SecretMissing", "MissingProviderSecret",
-		model.MissingRefStatus,
-		"UpdateFailed", "FailedScheduling",
-		"InvalidStoreConfiguration", "InvalidProviderConfig", "ValidationFailed":
-		return sevFailed
-	default:
-		if status == "" {
-			return sevBlank
-		}
-		return phraseSeverity(status)
-	}
-}
-
-// phraseSeverityWords maps lowercase words to a severity for statuses that miss
-// the exact-match table above — operators that put free-form phrases in
-// .status.phase (e.g. CloudNativePG's "Cluster in healthy state", "Failing
-// over") instead of conventional CamelCase values. Word-based, so "Unavailable"
-// or "Jumping" never match a fragment.
-var phraseSeverityWords = map[string]statusSev{
-	"failed": sevFailed, "failing": sevFailed, "error": sevFailed,
-	"errored": sevFailed, "degraded": sevFailed, "unhealthy": sevFailed,
-	"pending": sevProgressing, "waiting": sevProgressing, "creating": sevProgressing,
-	"initializing": sevProgressing, "starting": sevProgressing, "setting": sevProgressing,
-	"upgrading": sevProgressing, "provisioning": sevProgressing, "progressing": sevProgressing,
-	"progress": sevProgressing, "reconciling": sevProgressing, "restarting": sevProgressing,
-	"restarted": sevProgressing, "terminating": sevProgressing, "deleting": sevProgressing,
-	"healthy": sevRunning, "ready": sevRunning, "running": sevRunning, "active": sevRunning,
-	"succeeded": sevDone, "completed": sevDone,
-}
-
-// phraseSeverity classifies an unknown status string by scanning its words
-// against phraseSeverityWords. The worst-severity word wins ("Healthy but
-// degraded" is failed). A positive word preceded by "not" reads as
-// progressing-amber, mirroring the exact-match "NotReady". No recognized word
-// yields sevUnknown (gray).
-func phraseSeverity(status string) statusSev {
-	sev := sevUnknown
-	rank := func(s statusSev) int {
-		switch s {
-		case sevFailed:
-			return 3
-		case sevProgressing:
-			return 2
-		case sevRunning, sevDone:
-			return 1
-		default:
-			return 0
-		}
-	}
-	negated := false
-	for _, w := range strings.FieldsFunc(strings.ToLower(status), func(r rune) bool {
-		return !unicode.IsLetter(r) && !unicode.IsDigit(r)
-	}) {
-		if w == "not" {
-			negated = true
-			continue
-		}
-		if s, ok := phraseSeverityWords[w]; ok {
-			if negated && (s == sevRunning || s == sevDone) {
-				s = sevProgressing
-			}
-			if rank(s) > rank(sev) {
-				sev = s
-			}
-		}
-		negated = false
-	}
-	return sev
-}
-
-// StatusStyle returns the appropriate style for a resource status string.
-func StatusStyle(status string) lipgloss.Style {
-	switch statusSeverity(status) {
-	case sevDefault:
-		return lipgloss.NewStyle().Foreground(lipgloss.Color(ColorPrimary)).Background(BaseBg)
-	case sevRunning:
-		return StatusRunning
-	case sevDone:
-		return StatusOther
-	case sevProgressing:
-		return StatusProgressing
-	case sevNormal:
-		return DimStyle
-	case sevFailed:
-		return StatusFailed
-	case sevBlank:
-		return NormalStyle
-	default: // sevUnknown
-		return StatusOther
-	}
-}
-
-// StatusSeverityRank orders a status string worst-first (0 = most severe) for
-// status rollups and summaries. Derived from statusSeverity so it tracks
-// StatusStyle and works regardless of color profile.
-func StatusSeverityRank(status string) int {
-	switch statusSeverity(status) {
-	case sevFailed:
-		return 0
-	case sevProgressing:
-		return 2
-	case sevRunning:
-		return 3
-	default: // done / normal / default / blank / unknown
-		return 4
-	}
-}
-
-// StatusSortRank orders a status string healthy-first (0 = healthy) for the
-// Status column sort, where ascending has always put Running at the top.
-// Completed work ranks below both live and broken rows so a wall of Succeeded
-// cron pods never buries them. Derived from statusSeverity, so every status the
-// coloring understands gets a real bucket — the catch-all is reserved for rows
-// carrying no signal at all.
-func StatusSortRank(status string) int {
-	switch statusSeverity(status) {
-	case sevRunning:
-		return 0
-	case sevProgressing:
-		return 1
-	case sevFailed:
-		return 2
-	case sevDone:
-		return 3
-	default: // normal / default / blank / unknown
-		return 4
-	}
-}
-
-// condPolarity classifies how a status condition's type relates to health.
-type condPolarity int
-
-const (
-	condInfo    condPolarity = iota // neutral marker (Issuing, Progressing, Reconciling)
-	condReady                       // good when True, bad when False (Ready, Available, Synced)
-	condError                       // bad when True / present (Error, Failed, Degraded, Stalled)
-	condWarning                     // always amber (any "*Warning" type)
-)
-
-// conditionPolarities is a curated map of well-known custom-resource condition
-// types to their polarity. It makes the supported CRs explicit and overrides
-// the heuristic where the type name alone would misclassify (e.g. cert-manager
-// "Issuing", whose False state is normal, not a failure). Keys are lowercase.
-var conditionPolarities = map[string]condPolarity{
-	// ArgoCD Application (status-less, the type's presence is the signal).
-	"comparisonerror":         condError,
-	"invalidspecerror":        condError,
-	"unknownerror":            condError,
-	"syncerror":               condError,
-	"deletionerror":           condError,
-	"sharedresourcewarning":   condWarning,
-	"repeatedresourcewarning": condWarning,
-	"excludedresourcewarning": condWarning,
-	"orphanedresourcewarning": condWarning,
-	// cert-manager.
-	"ready":   condReady,
-	"issuing": condInfo, // actively issuing. False is the normal idle state
-	// external-secrets.
-	"secretsynced": condReady,
-	// FluxCD.
-	"reconciling": condInfo,
-	"stalled":     condError,
-	"healthy":     condReady,
-}
-
-var (
-	conditionErrorKeywords = []string{"error", "fail", "degrad", "invalid", "unhealthy", "pressure", "stalled", "lost", "backoff", "misconfig"}
-	conditionReadyKeywords = []string{"ready", "available", "synced", "healthy", "succeeded", "complete", "established", "approved", "provisioned", "bound", "scheduled", "initialized"}
-)
-
-// conditionPolarity returns the polarity of a condition type, consulting the
-// curated map first and falling back to a keyword heuristic.
-func conditionPolarity(condType string) condPolarity {
-	lower := strings.ToLower(condType)
-	if p, ok := conditionPolarities[lower]; ok {
-		return p
-	}
-	// Match keyword stems against whole camelCase tokens, not raw substrings,
-	// so negated types do not invert: "Unbound" -> ["unbound"] does not start
-	// with "bound", and "Incomplete" -> ["incomplete"] not with "complete".
-	tokens := conditionTokens(condType)
-	switch {
-	case tokenHasPrefix(tokens, []string{"warn"}):
-		return condWarning
-	case tokenHasPrefix(tokens, conditionErrorKeywords):
-		return condError
-	case tokenHasPrefix(tokens, conditionReadyKeywords):
-		return condReady
-	default:
-		return condInfo
-	}
-}
-
-// conditionTokens splits a PascalCase/camelCase condition type into lowercase
-// word tokens (e.g. "ContainersReady" -> ["containers", "ready"]). Runs of
-// non-letters and case transitions are token boundaries.
-func conditionTokens(condType string) []string {
-	var tokens []string
-	var cur []rune
-	flush := func() {
-		if len(cur) > 0 {
-			tokens = append(tokens, strings.ToLower(string(cur)))
-			cur = nil
-		}
-	}
-	for _, r := range condType {
-		switch {
-		case r >= 'A' && r <= 'Z':
-			flush()
-			cur = append(cur, r)
-		case (r >= 'a' && r <= 'z') || (r >= '0' && r <= '9'):
-			cur = append(cur, r)
-		default:
-			flush()
-		}
-	}
-	flush()
-	return tokens
-}
-
-// tokenHasPrefix reports whether any token starts with one of the keyword stems.
-func tokenHasPrefix(tokens, keywords []string) bool {
-	for _, tok := range tokens {
-		for _, kw := range keywords {
-			if strings.HasPrefix(tok, kw) {
-				return true
-			}
-		}
-	}
-	return false
-}
-
-// ConditionStyle returns the color style for a status condition, combining its
-// type's polarity with its status value. Conditions without a True/False status
-// (e.g. ArgoCD application conditions) are colored by type polarity alone.
-func ConditionStyle(condType, status string) lipgloss.Style {
-	if status == "Unknown" {
-		return DimStyle
-	}
-	switch conditionPolarity(condType) {
-	case condWarning:
-		return StatusWarning
-	case condError:
-		if status == "False" {
-			return StatusRunning // a failure condition that is False = healthy
-		}
-		return StatusFailed // True or status-less = problem
-	case condReady:
-		if status == "False" {
-			return StatusFailed // a readiness condition that is False = problem
-		}
-		return StatusRunning // True or status-less = healthy
-	default: // condInfo
-		if status == "True" {
-			return StatusProgressing // active / in progress
-		}
-		return DimStyle // False or status-less = neutral
 	}
 }
